Refactoring
Move Event building logic into a separate service.
